+++ dpytools/checks.py
@@ -20,10 +20,11 @@ Or inside Cogs:
 
 
 from discord.ext import commands
-from typing import Union
+from discord.ext.commands import PrivateMessageOnly
+from typing import Union, List
 from discord.utils import get
 
-from dpytools.errors import Unauthorized, IncorrectGuild
+from dpytools.errors import Unauthorized, IncorrectGuild, NotMemberOfCorrectGuild
 
 
 def admin_or_roles(*role_ids: Union[int, str]) -> commands.check:
@@ -73,9 +74,50 @@ def only_this_guild(guild_id: int) -> commands.check:
 
     async def predicate(ctx):
         if ctx.guild is None:
-            raise commands.NoPrivateMessage()
+            raise commands.NoPrivateMessage("Command was called from a direct message.")
         elif guild_id != ctx.guild.id:
-            raise IncorrectGuild()
+            raise IncorrectGuild(f"Command was called from a guild with id different than specified in check")
         return True
 
     return commands.check(predicate)
+
+
+def dm_from_this_guild(guild_id: int, delete: bool = False) -> commands.check:
+    """
+    The check returns true only if the specified :guild_id: is found within the bot's guilds.
+    For this check to work the guilds and members intents must be enabled.
+
+    Args:
+        guild_id: id of the guild the author of the context must be a part of for check to return True
+        delete: if True, it will try to delete ctx.message. This will only happen if the message is called
+                from the specificed guild_id, if the command is called from another guild the check will
+                just return False.
+            Possible exceptions by discord.py library:
+                commands.Forbidden – You do not have proper permissions to delete the message.
+                commands.NotFound – The message was deleted already
+                commands.HTTPException – Deleting the message failed.
+
+    Returns:
+        commands.check
+
+    Raises:
+        commands.PrivateMessageOnly: if called from a guild
+        dpytools.errors.NotMemberOfCorrectGuild: if not a member of the specified guild
+    """
+    async def predicate(ctx):
+        if ctx.guild and ctx.guild.id == guild_id:
+            if delete is True:
+                await ctx.message.delete()
+            raise PrivateMessageOnly("Command was called from the guild instead of a direct message.")
+        elif ctx.guild:
+            return False
+
+        guild = ctx.bot.get_guild(guild_id)
+
+        if ctx.author in guild.members:
+            return True
+        else:
+            raise NotMemberOfCorrectGuild("This command is unavailable to you.")
+
+    return commands.check(predicate)
+