+++ dpytools/errors.py
@@ -5,6 +5,9 @@ This Module contains all custom errors used in the package.
 
 from discord.ext.commands import CommandError
 
+class NotMemberOfCorrectGuild(CommandError):
+    pass
+
 
 class IncorrectGuild(CommandError):
     pass