+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.0.12a1
+- checks.dm_from_this_guild now accepts aditional argument :delete: to remove the message if called from a guild. This will work only if it comes from the specified guild. If its a different guild the check will just return False
+
+
+# 0.0.12a
+- added check `checks.dm_from_this_guild` 
 
 # 0.0.11a
 - Improved cogs command from owner_cog